:recycle: Remove account manager component and simplify feature complexity
