+++ src/modules/Auth/AccountsManagerScreen/index.js
@@ -1,14 +1,17 @@
 import React from 'react';
+import { useNavigation } from '@react-navigation/native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import { useTheme } from 'hooks/useTheme';
-import AccountsManager from 'modules/Accounts/components/AccountsManager';
+import AccountList from 'modules/Accounts/components/AccountList';
 
 import Splash from '../components/splash';
 
 import getAccountsManagerScreenStyles from './styles';
 
 export default function AccountsManagerScreen() {
+  const navigation = useNavigation();
+
   const { styles } = useTheme({
     styles: getAccountsManagerScreenStyles(),
   });
@@ -17,7 +20,15 @@ export default function AccountsManagerScreen() {
     <SafeAreaView style={[styles.wrapper, styles.theme.wrapper]}>
       <Splash animate={false} />
 
-      <AccountsManager />
+      <AccountList
+        mode="screen"
+        onEditAccountClick={(account) =>
+          navigation.navigate({ name: 'EditAccount', params: account })
+        }
+        onDeleteAccountClick={(account) =>
+          navigation.navigate({ name: 'DeleteAccount', params: account })
+        }
+      />
     </SafeAreaView>
   );
 }
